requirements.txt
flask
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
sentence-transformers
chromadb
llama-cpp-python
huggingface_hub
langchain_huggingface
langchain_chroma
unstructured
pdf2image
pytesseract
pypdf
flask_cors
python-magic
pdfminer.six
networkx
matplotlib
plotly
opensearch-py
python-dotenv
spacy